requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==0.0.6

chromadb==0.5.5
sentence-transformers==2.7.0
torch==2.3.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
fastapi==0.111.0
uvicorn==0.29.0
jinja2==3.1.4
transformers==4.41.2
tokenizers==0.19.1
